+++ packages/testing/src/steps/generic.impl.ts
@@ -125,6 +125,33 @@ export async function callWithMethodAndFourParams(
   }
 }
 
+export async function callWithMethodAndFiveParams(
+  world: PropsWorldLike,
+  field: string,
+  fnName: string,
+  param1: string,
+  param2: string,
+  param3: string,
+  param4: string,
+  param5: string
+): Promise<void> {
+  try {
+    const object = handleResolve(field, world) as Record<string, (...args: unknown[]) => unknown>;
+    const fn = object[fnName];
+    const result = await fn.call(
+      object,
+      handleResolve(param1, world),
+      handleResolve(param2, world),
+      handleResolve(param3, world),
+      handleResolve(param4, world),
+      handleResolve(param5, world)
+    );
+    world.props['result'] = result;
+  } catch (error) {
+    world.props['result'] = error;
+  }
+}
+
 export function referToAs(world: PropsWorldLike, from: string, to: string): void {
   world.props[to] = handleResolve(from, world);
 }