+++ toolbox/fdc3-for-web/demo/src/client/da/FDC3_2_1_JSONDirectory.ts
@@ -18,8 +18,8 @@ const convertToDirectoryList = (data: any): DirectoryApp[] => {
 };
 
 export class FDC3_2_1_JSONDirectory extends BasicDirectory {
-  constructor(fdc3Version = '3.0') {
-    super([], fdc3Version);
+  constructor(desktopAgentFdc3Version = '3.0') {
+    super([], desktopAgentFdc3Version);
   }
 
   async load(url: string) {